chore(styles): move interface wrappers to styles package
