+++ packages/styles/package.json
@@ -7,7 +7,8 @@
     "./accents/*": "./src/accents/*",
     "./partials/*": "./src/partials/*",
     "./theme/variables": "./src/theme/variables.ts",
-    "./theme/*": "./src/theme/*"
+    "./theme/*": "./src/theme/*",
+    "./wrappers": "./src/wrappers/index.ts"
   },
   "scripts": {
     "lint": "eslint --fix ./src",
